feat: Working with actions split
